Move methods to specific decorators when applicable
