+++ src/Sentry/Laravel/Features/Storage/CloudFilesystemDecorator.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Sentry\Laravel\Features\Storage;
+
+trait CloudFilesystemDecorator
+{
+    use FilesystemDecorator;
+
+    public function url($path)
+    {
+        return $this->withSentry(__FUNCTION__, func_get_args(), $path, compact('path'));
+    }
+}
